[enhancement] Add AbstractAdmins
