+++ openwisp_monitoring/device/admin.py
@@ -1,19 +1,19 @@
 from django.contrib import admin
 from django.urls import reverse
-from django.utils.html import format_html
-from django.utils.safestring import mark_safe
-from django.utils.translation import ugettext_lazy as _
 
-from openwisp_controller.config.admin import DeviceAdmin as BaseDeviceAdmin
 from openwisp_controller.config.models import Device
 
 from ..monitoring.admin import MetricAdmin
 from ..monitoring.models import Graph
 from . import settings as app_settings
 from .models import DeviceData, DeviceMonitoring
+from .base.admin import AbstractDeviceAdmin
+from .models import DeviceData
 
 
-class DeviceAdmin(BaseDeviceAdmin):
+class DeviceAdmin(AbstractDeviceAdmin):
+    model = Device
+
     def get_extra_context(self, pk=None):
         ctx = super(DeviceAdmin, self).get_extra_context(pk)
         if pk:
@@ -28,25 +28,6 @@ class DeviceAdmin(BaseDeviceAdmin):
                 }
             )
         return ctx
-
-    def health_status(self, obj):
-        return format_html(
-            mark_safe('<span class="health-{0}">{1}</span>'),
-            obj.monitoring.status,
-            obj.monitoring.get_status_display(),
-        )
-
-    health_status.short_description = _('health status')
-
-    def get_form(self, request, obj=None, **kwargs):
-        """
-        Adds the help_text of DeviceMonitoring.status field
-        """
-        health_status = DeviceMonitoring._meta.get_field('status').help_text
-        kwargs.update(
-            {'help_texts': {'health_status': health_status.replace('\n', '<br>')}}
-        )
-        return super().get_form(request, obj, **kwargs)
 
 
 DeviceAdmin.Media.js += MetricAdmin.Media.js + ('monitoring/js/percircle.js',)
